model/requirements.txt
torch
transformers
pillow
requests
optuna
scikit-learn
numpy
